+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Reduced database connection lock time in incoming activity queue executor.
+
 ## [4.5.1] - 2025-06-23
 
 ### Added